+++ signup.html
@@ -45,7 +45,9 @@
             </div>
           </div>
 
-          <button class="login-button flex-center" type="submit">회원가입</button>
+          <button class="login-button flex-center" type="submit">
+            회원가입
+          </button>
         </form>
       </div>
 
@@ -61,5 +63,6 @@
         </div>
       </div>
     </main>
+    <script src="js/signup.js" type="module"></script>
   </body>
 </html>
